+++ composeApp/src/commonMain/kotlin/ir/khebrati/audiosense/presentation/screens/result/ResultScreen.kt
@@ -1,27 +1,40 @@
 package ir.khebrati.audiosense.presentation.screens.result
 
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material3.CircularWavyProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.min
+import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import ir.khebrati.audiosense.domain.model.AcousticConstants
 import ir.khebrati.audiosense.presentation.components.AudiosenseScaffold
 import ir.khebrati.audiosense.presentation.navigation.AudiosenseRoute.DescriptiveResultRoute
 import ir.khebrati.audiosense.presentation.navigation.AudiosenseRoute.HomeRoute
 import ir.khebrati.audiosense.presentation.navigation.AudiosenseRoute.ResultRoute
 import ir.khebrati.audiosense.presentation.screens.result.TestResultUiState.*
-import ir.khebrati.audiosense.presentation.screens.result.components.Audiogram
 import ir.khebrati.audiosense.presentation.screens.result.components.DoneButton
 import ir.khebrati.audiosense.presentation.screens.result.components.HearingLossCard
 import ir.khebrati.audiosense.presentation.theme.AppTheme
@@ -78,11 +91,11 @@ private fun ReadyResultScreenContent(
             )
         }
         Spacer(modifier = Modifier.height(30.dp))
-        Audiogram(
-            leftAC = uiState.leftAC,
-            rightAC = uiState.rightAC,
-            modifier = Modifier.fillMaxWidth().height(300.dp),
-        )
+//        Audiogram(
+//            leftAC = uiState.leftAC,
+//            rightAC = uiState.rightAC,
+//            modifier = Modifier.fillMaxWidth().height(300.dp),
+//        )
         Spacer(modifier = Modifier.weight(1f))
         DoneButton(onClick = { onNavigateHome(HomeRoute) })
     }
